+++ Resources/config/routing.yml
@@ -12,7 +12,22 @@ KunstmaanAdminBundle_settings:
     resource: "@KunstmaanAdminBundle/Controller/SettingsController.php"
     type:     annotation
     prefix:   /admin/settings      
-    
+
+KunstmaanAdminBundle_user_settings:
+    resource: "@KunstmaanAdminBundle/Controller/UsersController.php"
+    type:     annotation
+    prefix:   /admin/settings/users
+
+KunstmaanAdminBundle_group_settings:
+    resource: "@KunstmaanAdminBundle/Controller/GroupsController.php"
+    type:     annotation
+    prefix:   /admin/settings/groups
+
+KunstmaanAdminBundle_role_settings:
+    resource: "@KunstmaanAdminBundle/Controller/RolesController.php"
+    type:     annotation
+    prefix:   /admin/settings/roles
+
 ###########################
 ## fos_userbundle routes ##
 ###########################
